MQTT/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
ca-certificates=2022.07.19=haa95532_0
certifi=2022.9.24=py38haa95532_0
openssl=1.1.1q=h2bbff1b_0
paho-mqtt=1.6.1=pypi_0
pip=22.2.2=py38haa95532_0
python=3.8.13=h6244533_0
setuptools=63.4.1=py38haa95532_0
sqlite=3.39.3=h2bbff1b_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
